requirements.txt
streamlit>=1.28.0
ultralytics>=8.0.0
opencv-contrib-python>=4.12.0
pillow>=9.5.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.15.0
streamlit-webrtc>=0.47.0
av>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
